log out, split up drawer
