+++ frontend-admin/src/components/common/drawer/style.ts
@@ -1,0 +1,10 @@
+const drawerWidth = 200;
+
+export const style = {
+    width: drawerWidth,
+    flexShrink: 0,
+    '& .MuiDrawer-paper': {
+        width: drawerWidth,
+        boxSizing: 'border-box',
+    },
+}
